Closes #3: Fix bug when using the same address more than once
The SOAP client in PHP creates XML references for identical elements which
cannot be disabled. Thus, when a user wants to order more than one stamp
and reuses and existing `AddressBinding` (for example because the sender
is the same for multiple stamps, which is quite common), the element is
only inserted into the XML once, with all further occurrences being
referenced by their ID (see #3 and https://stackoverflow.com/q/9744442).

The problem is that the DPAG web service doesn't expect (or accept) such
messages, returning an error similar to the following depending on the
element occuring multiple times:

`org.xml.sax.SAXParseException; cvc-complex-type.3.2.2: Attribute 'id'
is not allowed to appear in element 'ns1:address'`

Minimum reproducible example:

```
$partner_info = new \baltpeter\Internetmarke\PartnerInformation('ABCDE',
    1, 'yoursecretkey');
$service = new \baltpeter\Internetmarke\Service($partner_info);

$user_token = $service->authenticateUser('portokasse@yourmailserver.tld',
    'yourpassword')->getUserToken();

$address = new \baltpeter\Internetmarke\NamedAddress(new
    \baltpeter\Internetmarke\Name(new \baltpeter\Internetmarke\PersonName(
    'Salutation', 'Titel', 'Kim', 'Mustermensch'), null),
    new \baltpeter\Internetmarke\Address('', 'Musterstraße', '123',
    '12345', 'Musterstadt', 'DEU'));
$addresses = new \baltpeter\Internetmarke\AddressBinding($address,
    $address);
$order_item = new \baltpeter\Internetmarke\OrderItem(11, 255,
    $addresses, new \baltpeter\Internetmarke\Position(1, 5, 1),
    'AddressZone');

$service->checkoutShoppingCartPdf($user_token, 1, array($order_item,
    $order_item), 2 * 85, null, null, null, null);
```

I was unfortunately not able to find an elegant solution to this
problem. As far as I am aware, this behaviour cannot be turned off
in the PHP SOAP client.
Thus, the workaround I used is to create a deep copy (using the
DeepCopy library by My C-Labs, https://github.com/myclabs/DeepCopy).
The copy is made in the setters. To make sure a similar issue cannot
occur in the future, I have applied this procedure to all case where
a non-primitive value is assigned through a setter. In addition, I
have transitioned all constructors in the classes to use the setters
instead of direct assignments for consistency and easier workarounds
like this in the future.